Renamed from NLog.Redis to NLog.Targets.Redis
