--- src/NLog.Redis/NLog.Redis.csproj
@@ -1,30 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net45;netstandard1.5</TargetFrameworks>
-    <Title>NLog Target for Redis</Title>
-    <Description>NLog Target for Redis</Description>
-    <copyright>Copyright 2014</copyright>
-    <Authors>Rich Clement</Authors>
-
-    <Version>1.0.3</Version>
-    <PackageReleaseNotes>Added support for NetStandard1.5</PackageReleaseNotes>
-    <PackageTags>NLog;Redis;Log;Logging</PackageTags>
-    <PackageProjectUrl>https://github.com/richclement/NLog.Redis</PackageProjectUrl>
-    <PackageLicenseUrl>https://raw.githubusercontent.com/richclement/NLog.Redis/master/LICENSE</PackageLicenseUrl>
-    <RepositoryType>git</RepositoryType>
-    <RepositoryUrl>git://github.com/richclement/NLog.Redis</RepositoryUrl>
-    
-  </PropertyGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net45' ">
-    <PackageReference Include="NLog" Version="4.0.0" />
-    <PackageReference Include="StackExchange.Redis" Version="1.0.488" />
-  </ItemGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard1.5' ">
-    <PackageReference Include="NLog" Version="4.5.0" />
-    <PackageReference Include="StackExchange.Redis" Version="1.1.605" />
-  </ItemGroup>
-
-</Project>
