+++ src/NLog.Targets.Redis.Tests/RedisTargetTests.cs
@@ -1,11 +1,10 @@
 ﻿using NLog.Config;
-using NLog.Redis.Tests.Mocks;
-using NLog.Targets;
+using NLog.Targets.Redis.Tests.Mocks;
 using NSubstitute;
 using StackExchange.Redis;
 using Xunit;
 
-namespace NLog.Redis.Tests
+namespace NLog.Targets.Redis.Tests
 {
     public class RedisTargetTests
     {
