+++ src/NLog.Targets.Redis.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28010.2026
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.Targets.Redis", "NLog.Targets.Redis\NLog.Targets.Redis.csproj", "{D30CC839-170F-4B66-A075-A0EB3FA31FC6}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NLog.Targets.Redis.Tests", "NLog.Targets.Redis.Tests\NLog.Targets.Redis.Tests.csproj", "{6781DE54-6720-4487-9B7C-DD9C54EFF19A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D30CC839-170F-4B66-A075-A0EB3FA31FC6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D30CC839-170F-4B66-A075-A0EB3FA31FC6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D30CC839-170F-4B66-A075-A0EB3FA31FC6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D30CC839-170F-4B66-A075-A0EB3FA31FC6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6781DE54-6720-4487-9B7C-DD9C54EFF19A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6781DE54-6720-4487-9B7C-DD9C54EFF19A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6781DE54-6720-4487-9B7C-DD9C54EFF19A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6781DE54-6720-4487-9B7C-DD9C54EFF19A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {02C7646E-CC7A-4177-8958-198E1A8FDC5C}
+	EndGlobalSection
+EndGlobal
